backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
llama-cpp-python==0.3.16
huggingface-hub==0.20.1
python-dotenv==1.0.0
